+++ src/interfaces/settings/index.ts
@@ -30,3 +30,4 @@ export { iLedSettings } from './iLedSettings';
 export { iSceneSettings } from './iSceneSettings';
 export { iSettingsProvider } from './iSettingsProvider';
 export { iWledSettings } from './iWledSettings';
+export { iGlobalHeaterSettings } from './iGlobalHeaterSettings';